backend/requirements.txt
fastapi==0.87.0
uvicorn==0.19.0
motor==3.0.0
pydantic==1.10.2
pymongo==4.3.2
python-dotenv==0.21.0
bson
pydantic[email]
python-jose==3.3.0
passlib==1.7.4
bcrypt==3.2.2
python-multipart==0.0.5